gsicrawler/requirements.txt
celery==4.1.1
sqlalchemy
redis
Flask
requests
connexion[swagger-ui]
flower
connexion
elasticsearch
rdflib
rdflib-jsonld 
newspaper3k
tweepy
beautifulsoup4
unidecode
newsapi-python
nytimesarticle
tornado